Split Util.HTTP2.Stream module
... into three modules: common, server and client
The common part is HTTP2 independent.
