+++ grapesy/src/Network/GRPC/Util/Session/Client.hs
@@ -20,7 +20,8 @@ import Network.HTTP.Types qualified as HTTP
 import Network.HTTP2.Client qualified as Client
 
 import Network.GRPC.Util.HTTP2 (fromHeaderTable)
-import Network.GRPC.Util.HTTP2.Stream
+import Network.GRPC.Util.HTTP2.ClientStream
+import Network.GRPC.Util.Stream
 import Network.GRPC.Util.RedundantConstraint (addConstraint)
 import Network.GRPC.Util.Session.API
 import Network.GRPC.Util.Session.Channel
